// src/seed/tenantSpecific/featureConfiguration/features/CreateJobOrderForm/enum/createJobOrderFormCellInputName.enum.ts
export enum CreateJobOrderFormCellInputName {
  Name = 'name',
  Location = 'location',
  Branch = 'branch',
  ContractType = 'contractType',
  StartDate = 'startDate',
  EndDate = 'endDate',
  StartTime = 'startTime',
  EndTime = 'endTime',
  NumberOfOpenings = 'numberOfOpenings',
  Salary = 'salary',
  SalaryHigh = 'salaryHigh',
  Rate = 'rate',
  Shifts = 'shifts',
  Days = 'days',
  Role = 'role',
  Languages = 'languages',
  Certifications = 'certifications',
  JobDescription = 'jobDescription',
  AdditionalInformation = 'additionalInformation',
  InterviewRequired = 'interviewRequired',
  DateRange = 'dateRange',
  DateStart = 'dateStart',
  Experience = 'experience',
  TimePicker = 'timePicker',
  DayOneGuidance = 'dayOneGuidance',
  SectorLevelId = 'sectorLevelId',
  Sector = 'sector',
  EmploymentType = 'employmentType',
  WorkType = 'workType',
  TimeSheetApprover = 'timeSheetApprover',
  ReportTo = 'reportTo',
  BillTo = 'billTo',
  AttachedFiles = 'files',
}